Remove LIB_JSONNET_VERSION from libjsonnet.h
It could be very confusing if header is linked with older/newer
version of the library. We have jsonnet_version function which
returns the version of the actual library underneath.

So unless it's some standard thing to provide the version of the
header separate from the version of the actual library, I think it's
just a bad idea to include it, especially considering go-jsonnet C bindings.