Change sort algorithm
Current std.sort implementation recurses may recurse N
times for the array of size N. For example for sorted or
reverse sorted data. This causes problems with std.set for example.

This change implements mergesort which always divides the array
in half, so the depth of recursion is always reasonable.

For smaller arrays quicksort is still used (it is faster).