cmake: Use CMAKE_CURRENT_SOURCE_DIR instead of CMAKE_SOURCE_DIR
When Jsonnet CMake project is included in another project using
add_subdirectory, CMAKE_SOURCE_DIR points to the root directory of the
main project but not the root of the Jsonnet tree.
To avoid this problem, instead use CMAKE_CURRENT_SOURCE_DIR, which
points to the directory where the current CMakeList resides.