python: use CMake, make or distutils compiler for libjsonnet build
This change takes advantage of distutils compiler feature to compile libjsonnet when neither CMake and make are available, or when both CMake and make build are failing.

A fallback strategy to go through this chain of tools (CMake, make, distutils compiler).

This also port stdlib/to_c_array.cpp to pure python to generate std.jsonnet.h.

Python extension can now be build and installed properly on windows and Python 3.5+ when Visual C++ Build Tools are installed.

Windows build on Python 2.7 is still not supported as this version relies on a old version of Visual C++ tools that doesn't support c++11.