Add language reference
This is not a complete reference yet, but it
should be a solid base. It covers about ~50%
of what I think it should ideally cover, but
we have to start somewhere.

I purposefully chose a fairly conversational,
explanatory tone (for a language reference).
We already have a more formal spec, so this
is aimed more at users who want to understand
some concepts better than at experts, who
are interested in boundary cases and I think
a casual tone is easier to read for a casual
user. Also, it is much easier for me to write
that way.