Array comparison and sourceVals.
This change introduces the ability to compare
arrays lexicographically using the customary operator (<, >, <=, >=).

This change needs to be followed up by appropriate changes in the docs
and go-jsonnet implementation.

To make this change, I introduced "sourceVals" – the values from std
which can be used relatively easily in the interpreter. In particular
the operator< is implemented as jsonnet function __array_less which
is called by the interpreter.

This is much less efficient than a builtin implementation of course,
but implmentation of builtins here tends to be really painful and
this allows us to quickly implement features which would otherwise
be very tricky. Also if performance is needed, go-jsonent should be used
anyway (where builtin implemetation requires easily 10x less effort).