More flexibile overriding of test results.
We allow implementations to have different error
outputs (for very good reasons). We still want
to share the test suite. So far we had an "implementation"
suffix. This still required all the variants to be put
in this repo. This is inconvenient, because we need to
sync repos on every change which changes stack traces.

This change allows providing an override directory
with alternative test results. In addition to making
work on go-jsonnet easier, it will also allow 3rd party
implementations to use the test suite directly.