Add a missing header.
Some users reported problems with compiling the project. It seems like on some platforms we need to explicitly include a header which is not needed for the ones we test on. I have no way to reproduce, but hopefully that will fix it.

Thanks to @prazek for the tip about what might be wrong.