Change com_google_googletest to a non-dev dependency (#1090)
I had to patch the MODULE.bazel file as part of publishing to BCR (https://github.com/bazelbuild/bazel-central-registry/pull/667), as some targets failed to build when googletest was marked as a dev dependency.
I think it would be possible to get around this by limiting which parts of this repo are built by the BCR presubmit checks. Making googletest a normal dependency seems like the better option, as it will allow the presubmit checks to be more thorough.