fix fmt stdlib.jsonnet test case in test_suite
jsonnetfmt (by default) rewrites string literals, unescaping and
re-escaping them. This changes which characters are escaped, and in
particular the output has character \u00A0 (Non-breaking space) in its
raw form, not as an escape sequence. However the stdlib.jsonnet test
case includes this character as an escaped sequence in a string literal,
so it doesn't pass unchanged through jsonnetfmt - this causes the fmt
tests to fail.

Since putting a raw non-breaking space character directly into the
source is confusing / ambiguous (since it just looks like a normal space
character), this change instead uses a std.char() call to generate a
U+00A0 character at runtime, so that it is both unambiguous in the
source and passes unchanged through jsonnetfmt.