add GitHub workflow to build and test all PRs and pushes
There are three jobs in the workflow (jobs run independently)
  - Build and Test using the plain Makefile and ./tests.sh script.
  - Build and Test using CMake & CTest.
  - Check code formatting using clang-format.

The last (check formatting) is disabled because right now the code
doesn't conform to clang-format's desired output. Once the code has
been reformatted, the workflow job can be enabled so that future PRs
can be automatically checked for format regressions.