Makefile: switch to using -Werror for test-formatting
Instead of running clang-format in a mode that generates XML and then
grepping and counting that output and using the shell 'test' command
on _that_ output to work out whether the formatter recommended changes,
use clang-format --dry-run -Werror so that the formatter will exit with
a failure code if it finds anything it wants to reformat, and not
otherwise do anything.

This just makes the 'test-formatting' make target a little simpler.