change setup.py to build directly not using the Makefile
The Makefile is problematic because it needs shell and the `od` command.
But Python setuptools is perfectly capable of building some .cpp files
on its own, so we can use that, it just needs a minor extension to
generate core/std.jsonnet.h.

Also update MANIFEST.in after the recent changes to third_party:
The bundled RapidYAML was switched to the single-header version, and
nlohmann json.hpp was moved to a subdirectory.