Clamp negative slice start range to prevent out-of-range, and update fmt test golden
Slice is documented to behave similarly to Python. Negative indexes in
Python slices are clamped so slice indexes never produce out-of-range
runtime errors. Update the 'slice' sugar to do the same.

Note that only the starting index needs to be clamped in this way; the
ending index can be out-of-range doesn't cause an error, as the loop
exit condition compares the current index against the ending index.