fix RapidYAML build errors when using system provided json.hpp
When building with bundled JSON and bundled RapidYAML, the build works.
However, when building with system-provided JSON and bundled RapidYAML, the build
fails due to many things not being defined. It seems these are supposed to come
from the standard <charconv> header, but RapidYAML 0.5.0 release erroneously
strips that include in its single-file amalgamation. See RapidYAML issue

https://github.com/biojppm/rapidyaml/issues/364#issuecomment-1536625415