chore: define a local jsonnet_json_golden_test rule instead of using rules_jsonnet
Although rules_jsonnet is the appropriate Bazel module for external
Jsonnet users to use for their Jsonnet build actions, using it
internally within the jsonnet repo itself is actually a little difficult
as it creates a mutually recursive dependency.

Since the only thing it's being used for is the jsonnet_json_test rule,
and we have scripts to runs such tests anyway, it's relatively easy to
define our own jsonnet_json_golden_test rule and avoid the recursive
dependency.