doc/man: Add man pages jsonnet.1 and jsonnetfmt.1
This commit adds man pages for the jsonnet and jsonnetfmt commands.
The Debian Lintian quality checker warns that each binary should have
a manual page, but hopefully this idea would also apply to non-Debian users.
These man pages are created using the help2man command, so the contents are
substantially the same as the output of the help option.