Fixes build on CMake 4
This patch resolves an issue where the project fails to build with newer versions of CMake,
specifically CMake 4.0, which has removed backward compatibility with versions older than 3.5.

The bug was originally reported on the Debian issue tracker at #1113063:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1113063

The project's CMakeLists.txt and CMakeLists.txt.in files specify a cmake_minimum_required version of 2.8.7...4.0.
This causes the build process to terminate with a fatal error on CMake 4.0 or later.

The fix updates the required minimum version to 3.5...4.1 to align with the latest CMake policies
and allow the project to build successfully.
This change will be beneficial not only for Debian but for all other Linux distributions
and operating systems running a modern CMake version.

Signed-off-by: Fukui Daichi <a.dog.will.talk@akane.waseda.jp>